@@ -1,6 +1,6 @@
 module github.com/smartforce-io/atc
 
-go 1.16
+go 1.23.5
 
 require (
 	github.com/dgrijalva/jwt-go v3.2.0+incompatible
@@ -8,4 +8,13 @@ require (
 	github.com/gorilla/mux v1.8.0
 	golang.org/x/oauth2 v0.0.0-20210819190943-2bc19b11175f
 	gopkg.in/yaml.v2 v2.4.0
+)
+
+require (
+	github.com/golang/protobuf v1.4.2 // indirect
+	github.com/google/go-querystring v1.1.0 // indirect
+	golang.org/x/crypto v0.0.0-20210817164053-32db794688a5 // indirect
+	golang.org/x/net v0.0.0-20210226172049-e18ecbb05110 // indirect
+	google.golang.org/appengine v1.6.7 // indirect
+	google.golang.org/protobuf v1.25.0 // indirect
 )
